requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-community>=0.0.10
langgraph>=0.0.20
faiss-cpu>=1.7.4
chromadb>=0.4.18
openai>=1.6.0
python-dotenv>=1.0.0
pandas>=2.2.0
numpy>=1.24.0
requests>=2.31.0
beautifulsoup4>=4.12.0
tiktoken>=0.5.0
streamlit-chat>=0.1.0
plotly>=5.17.0 